requirements.txt
APScheduler==3.2.0
Flask==0.11.1
Flask-SQLAlchemy==2.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.14
Werkzeug==0.11.10
apiai==1.0.2
argparse==1.2.1
click==6.6
funcsigs==1.0.2
futures==3.0.5
gunicorn==19.6.0
itsdangerous==0.24
numpy==1.11.1
pytz==2016.6.1
requests==2.10.0
schedule==0.3.2
six==1.10.0
tzlocal==1.2.2
wsgiref==0.1.2
psycopg2==2.6.1
whitenoise==2.0.6
sqlalchemy-migrate==0.9.2
